Code Optimization
Made Changes in Code
- Added modules
- Image Path conflict
- Made changes in code to better understanding
