+++ Dockerfile
@@ -7,4 +7,6 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 COPY ./src ./src
 
-CMD ["flask", "--app", "src.main", "run", "--host=0.0.0.0", "--port=80"]
+WORKDIR /code/src
+
+CMD ["flask", "--app", "main", "run", "--host=0.0.0.0", "--port=80"]
