--- docker-compose.yml
@@ -1,12 +1,0 @@
-version: "1"
-
-services:
-  web:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    container_name: web
-    ports:
-      - "80:3000"
-    volumes:
-      - C:\Users\amrit\Documents\CapStone\DecorGene:/src:ro
